Add class-based HTTP client
