+++ packages/http/src/stamp.webcrypto.ts
@@ -1,9 +1,9 @@
 /// <reference lib="dom" />
 
 import { uint8ArrayToHexString, convertTurnkeyApiKeyToJwk } from "./encoding";
-import type { TStamper } from "./shared";
+import type { TLegacyStamper } from "./shared";
 
-export const stamp: TStamper = async (input: {
+export const stamp: TLegacyStamper = async (input: {
   content: string;
   publicKey: string;
   privateKey: string;
